Imagens e Curso carregando no formulario editar

+++ src/app/project/project.component.ts
@@ -1,4 +1,4 @@
-import { DialogService } from './../dialog.service';
+import { DialogService } from '../services/dialog.service';
 import { CourseService } from './../services/course.service';
 import { ProjectFormComponent } from './project-form/project-form.component';
 import { MatDialogConfig } from '@angular/material/dialog';
@@ -6,11 +6,10 @@ import { MatDialog } from '@angular/material/dialog';
 import { FormModelService } from './../services/form-model.service';
 import "@angular/compiler";
 import { ResponseService } from './../services/response.service';
-import { ActivatedRoute, Params, Router } from '@angular/router';
 import { ProjectService } from './../services/project.service';
 import { Project } from './../models/project.model';
-import { Component, Input, OnInit } from '@angular/core';
-import {MatTableDataSource} from '@angular/material/table';
+import { Component, OnInit } from '@angular/core';
+import { MatTableDataSource } from '@angular/material/table';
 import { Course } from '../models/course.model';
 
 @Component({
@@ -18,96 +17,94 @@ import { Course } from '../models/course.model';
   templateUrl: './project.component.html',
   styleUrls: ['./project.component.css']
 })
-export class ProjectComponent implements OnInit{
-  
-  courseList :Course[] = [];
-  ELEMENG_DATA :Project[] = [];
-  displayedColumns :string[] = ['name', 'image', 'why', 'what', 'whatWillWeDo', 'projectStatus', 'course', 'actions'];
+export class ProjectComponent implements OnInit {
+
+  courseList: Course[] = [];
+  ELEMENG_DATA: Project[] = [];
+  displayedColumns: string[] = ['name', 'image', 'why', 'what', 'whatWillWeDo', 'projectStatus', 'course', 'actions'];
   dataSource = new MatTableDataSource<Project>(this.ELEMENG_DATA);
   searchKey!: string;
-  
-  constructor(private projectService :ProjectService,
-              private courseService :CourseService,
-              private service :FormModelService,
-              private dialog :MatDialog,
-              private dialogService :DialogService,
-              private responseService :ResponseService,
-              private router :Router,
-              private route :ActivatedRoute) { }
 
-  ngOnInit(): void {   
-    this.loadProjects();     
-    this.loadCourses(); 
+  constructor(private projectService: ProjectService,
+    private courseService: CourseService,
+    private service: FormModelService,
+    private dialog: MatDialog,
+    private dialogService: DialogService,
+    private responseService: ResponseService) { }
+
+  ngOnInit(): void {
+    this.loadProjects();
+    this.loadCourses();
   }
-  
-  loadProjects(){
+
+  loadProjects() {
     this.projectService.getAll().subscribe(
-      (query :Project[]) => {
+      (query: Project[]) => {
         this.dataSource.data = query as Project[];
       },
-       (message :any) => {
-         console.error(message);
-       });
+      (message: any) => {
+        console.error(message);
+      });
   }
-  
-  loadCourses(){
+
+  loadCourses() {
     this.courseService.getAll().subscribe(
       data => {
         this.courseList = data;
       }
     )
   }
-  
-  onGetByCourse(id :number){
+
+  onGetByCourse(id: number) {
     this.projectService.getByCourse(id).subscribe(
-      (query :Project[]) => {
+      (query: Project[]) => {
         this.dataSource.data = query as Project[];
       },
-       (message :any) => {
-         console.error(message);
-       });
+      (message: any) => {
+        console.error(message);
+      });
   }
-  
-  verifyCourseId(id :number){
-    if(id == 0){
+
+  verifyCourseId(id: number) {
+    if (id == 0) {
       return
     }
     this.onGetByCourse(id)
   }
-  
+
   onSearchClear() {
     this.searchKey = "";
     this.applyFilter();
   }
-  
+
   applyFilter() {
     this.dataSource.filter = this.searchKey.trim().toLowerCase();
   }
-  
-  onCreate(){
+
+  onCreate() {
     this.service.initializeFormGroup();
     const dialogConfig = new MatDialogConfig();
     dialogConfig.disableClose = true;
     dialogConfig.autoFocus = true;
     dialogConfig.width = "40%";
-    this.dialog.open(ProjectFormComponent,dialogConfig);
+    this.dialog.open(ProjectFormComponent, dialogConfig);
     this.dialog.afterAllClosed.subscribe(() => this.loadProjects());
   }
-  
-  onEdit(project :Project){
+
+  onEdit(project: Project) {
     this.service.populateForm(project);
     const dialogConfig = new MatDialogConfig();
     dialogConfig.disableClose = true;
     dialogConfig.autoFocus = true;
     dialogConfig.width = "40%";
-    this.dialog.open(ProjectFormComponent,dialogConfig);
+    this.dialog.open(ProjectFormComponent, dialogConfig);
     this.dialog.afterAllClosed.subscribe(() => this.loadProjects());
   }
-    
-  onDelete(id :number){
-      this.dialogService.openConfirmDialog('Você tem certeza que deseja apagar este projeto ?')
+
+  onDelete(id: number) {
+    this.dialogService.openConfirmDialog('Você tem certeza que deseja apagar este projeto ?')
       .afterClosed().subscribe(response => {
-        if(response){
+        if (response) {
           this.projectService.delete(id).subscribe(
             data => {
               console.log(data);
@@ -120,6 +117,6 @@ export class ProjectComponent implements OnInit{
           )
         }
       })
-  } 
-    
+  }
+
 }
